% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ev_means_equal}
\alias{ev_means_equal}
\title{Compute expected value as function of n, N}
\usage{
ev_means_equal(n, N, means, sd, sig.level, ties)
}
\arguments{
\item{n}{Sample size per group}

\item{N}{Population size (estimate)}

\item{means}{Vector of estimated means}

\item{sd}{Standard deviation of the groups (assumed equal)}

\item{sig.level}{Significance level}

\item{ties}{Tie-breaking probability}
}
\value{
A scalar indicating the expected mean reward per unit in the population
}
\description{
Comparing means with equal variances
}
