% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ev_means_unequal}
\alias{ev_means_unequal}
\title{Compute expected value as function of n, N}
\usage{
ev_means_unequal(n, N, means, sds, sig.level, ties)
}
\arguments{
\item{n}{Sample size per group}

\item{N}{Population size (estimate)}

\item{means}{Vector of estimated means}

\item{sds}{Vector of standard deviation of the groups}

\item{sig.level}{Significance level}

\item{ties}{Tie-breaking probability}
}
\value{
A scalar indicating the expected mean reward per unit in the population
}
\description{
Comparing means with unequal variances
}
