% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ev_proportions}
\alias{ev_proportions}
\title{Compute expected value as function of n, N}
\usage{
ev_proportions(n, N, proportions, sig.level, ties)
}
\arguments{
\item{n}{Sample size per group}

\item{N}{Population size (estimate)}

\item{proportions}{Vector of two proportions}

\item{sig.level}{Significance level}

\item{ties}{Tie-breaking probability}
}
\value{
A scalar indicating the expected mean reward per unit in the population
}
\description{
Comparing proportions
}
