% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.ssev}
\alias{print.ssev}
\title{Pretty printing of ssev object}
\usage{
\method{print}{ssev}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{Object of type ssev for pretty printing}

\item{digits}{Standard number of digits for pretty printing, default is getOption("digits")}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Pretty printing of ssev object
}
